+++ modules/elementary_functions/builtin/cpp/Gateway.cpp
@@ -188,8 +188,8 @@ static const nlsGateway gateway[] = {
     { "sort", Nelson::ElementaryFunctionsGateway::sortBuiltin, 2, -1 },
     { "linspace", Nelson::ElementaryFunctionsGateway::linspaceBuiltin, 1, 3 },
     { "log2", Nelson::ElementaryFunctionsGateway::log2Builtin, 2, 1 },
-    { "max", Nelson::ElementaryFunctionsGateway::maxBuiltin, 2, 3 },
-    { "min", Nelson::ElementaryFunctionsGateway::minBuiltin, 2, 3 },
+    { "max", Nelson::ElementaryFunctionsGateway::maxBuiltin, 2, 4 },
+    { "min", Nelson::ElementaryFunctionsGateway::minBuiltin, 2, 4 },
     { "fliplr", Nelson::ElementaryFunctionsGateway::fliplrBuiltin, 1, 1 },
     { "flipud", Nelson::ElementaryFunctionsGateway::flipudBuiltin, 1, 1 },
 };