+++ modules/elementary_functions/src/cpp/Maximum.cpp
@@ -33,36 +33,137 @@ namespace Nelson {
 //=============================================================================
 template <class T>
 void
-TMaxGreaterReal(
+TMaxLessInteger(
     const T* spx, const T* spy, T* dp, indexType count, indexType stridex, indexType stridey)
 {
 #if defined(_NLS_WITH_OPENMP)
 #pragma omp parallel for
 #endif
     for (ompIndexType i = 0; i < (ompIndexType)count; i++) {
-        dp[i] = (spx[stridex * i] > spy[stridey * i]) ? spx[stridex * i] : spy[stridey * i];
+        T x = spx[stridex * i];
+        T y = spy[stridey * i];
+        dp[i] = (x > y) ? x : y;
     }
 }
 //=============================================================================
 template <class T>
 void
-TMaxGreaterComplex(
-    const T* spx, const T* spy, T* dp, indexType count, indexType stridex, indexType stridey)
+TMaxLessReal(bool omitnan, const T* spx, const T* spy, T* dp, indexType count, indexType stridex,
+    indexType stridey)
 {
+    if (omitnan) {
 #if defined(_NLS_WITH_OPENMP)
 #pragma omp parallel for
 #endif
-    for (ompIndexType i = 0; i < (ompIndexType)count; i++) {
-        T xmag = complex_abs(spx[2 * stridex * i], spx[2 * stridex * i + 1]);
-        T ymag = complex_abs(spy[2 * stridey * i], spy[2 * stridey * i + 1]);
-        if (xmag > ymag) {
-            dp[2 * i] = spx[2 * stridex * i];
-            dp[2 * i + 1] = spx[2 * stridex * i + 1];
-        } else {
-            dp[2 * i] = spy[2 * stridey * i];
-            dp[2 * i + 1] = spy[2 * stridey * i + 1];
+        for (ompIndexType i = 0; i < (ompIndexType)count; i++) {
+            T x = spx[stridex * i];
+            T y = spy[stridey * i];
+            if (std::isnan(x) && std::isnan(y)) {
+                dp[i] = (T)std::nan("NaN");
+            } else if (std::isnan(x) && !std::isnan(y)) {
+                dp[i] = y;
+            } else if (std::isnan(y) && !std::isnan(x)) {
+                dp[i] = x;
+            } else {
+                dp[i] = (x > y) ? x : y;
+            }
+        }
+    } else {
+#if defined(_NLS_WITH_OPENMP)
+#pragma omp parallel for
+#endif
+        for (ompIndexType i = 0; i < (ompIndexType)count; i++) {
+            T x = spx[stridex * i];
+            T y = spy[stridey * i];
+            if (std::isnan(x) || std::isnan(y)) {
+                dp[i] = (T)std::nan("NaN");
+            } else {
+                dp[i] = (x > y) ? x : y;
+            }
         }
     }
+}
+//=============================================================================
+template <class T>
+void
+TMaxLessComplex(bool omitnan, const T* spx, const T* spy, T* dp, indexType count, indexType stridex,
+    indexType stridey)
+{
+    if (omitnan) {
+#if defined(_NLS_WITH_OPENMP)
+#pragma omp parallel for
+#endif
+        for (ompIndexType i = 0; i < (ompIndexType)count; i++) {
+            T xre = spx[2 * stridex * i];
+            T xim = spx[2 * stridex * i + 1];
+            T yre = spy[2 * stridey * i];
+            T yim = spy[2 * stridey * i + 1];
+
+            T xmag = complex_abs(xre, xim);
+            T ymag = complex_abs(yre, yim);
+            if (std::isnan(xmag) && std::isnan(ymag)) {
+                dp[2 * i] = xre;
+                dp[2 * i + 1] = xim;
+            } else if (std::isnan(xmag)) {
+                dp[2 * i] = yre;
+                dp[2 * i + 1] = yim;
+            } else if (std::isnan(ymag)) {
+                dp[2 * i] = xre;
+                dp[2 * i + 1] = xim;
+            } else {
+                if (xmag > ymag) {
+                    dp[2 * i] = xre;
+                    dp[2 * i + 1] = xim;
+                } else {
+                    dp[2 * i] = yre;
+                    dp[2 * i + 1] = yim;
+                }
+            }
+        }
+    } else {
+#if defined(_NLS_WITH_OPENMP)
+#pragma omp parallel for
+#endif
+        for (ompIndexType i = 0; i < (ompIndexType)count; i++) {
+            T xre = spx[2 * stridex * i];
+            T xim = spx[2 * stridex * i + 1];
+            T yre = spy[2 * stridey * i];
+            T yim = spy[2 * stridey * i + 1];
+            T xmag = complex_abs(xre, xim);
+            T ymag = complex_abs(yre, yim);
+            if (std::isnan(xmag) && std::isnan(ymag)) {
+                dp[2 * i] = xre;
+                dp[2 * i + 1] = xim;
+            } else if (std::isnan(xmag)) {
+                dp[2 * i] = xre;
+                dp[2 * i + 1] = xim;
+            } else if (std::isnan(ymag)) {
+                dp[2 * i] = yre;
+                dp[2 * i + 1] = yim;
+            } else {
+                if (xmag > ymag) {
+                    dp[2 * i] = xre;
+                    dp[2 * i + 1] = xim;
+                } else {
+                    dp[2 * i] = yre;
+                    dp[2 * i + 1] = yim;
+                }
+            }
+        }
+    }
+}
+//=============================================================================
+template <class T>
+void
+TMaxAllInteger(const T* sp, T* dp, indexType elementCount)
+{
+    T maxval = sp[0];
+    for (ompIndexType k = 1; k < (ompIndexType)elementCount; ++k) {
+        if (sp[k] > maxval) {
+            maxval = sp[k];
+        }
+    }
+    dp[0] = maxval;
 }
 //=============================================================================
 template <class T>
@@ -72,135 +173,288 @@ TMaxInteger(
 {
     T maxval;
     double maxdex;
-    if (planes == 1 && planesize == 1) {
-        maxval = sp[0];
-        maxdex = 0;
-#if defined(_NLS_WITH_OPENMP)
-#pragma omp parallel for ordered
-#endif
-        for (ompIndexType k = 0; k < (ompIndexType)linesize; k++) {
-            if (sp[k] > maxval) {
+    for (indexType i = 0; i < planes; i++) {
+        for (indexType j = 0; j < planesize; j++) {
+            maxval = sp[i * planesize * linesize + j];
+            maxdex = 0;
+            for (ompIndexType k = 0; k < (ompIndexType)linesize; k++) {
+                T val = sp[i * planesize * linesize + j + k * planesize];
+                if (val > maxval) {
+                    maxval = val;
+                    maxdex = (double)k;
+                }
+            }
+            dp[i * planesize + j] = maxval;
+            if (iptr != nullptr) {
+                iptr[i * planesize + j] = maxdex + 1;
+            }
+        }
+    }
+}
+//=============================================================================
+template <class T>
+void
+TMaxAllReal(bool omitnan, const T* sp, T* dp, indexType elementCount)
+{
+    if (omitnan) {
+        T maxval;
+        bool init = false;
+        for (ompIndexType k = 0; k < (ompIndexType)elementCount; ++k) {
+            if (!std::isnan(sp[k])) {
                 maxval = sp[k];
-                maxdex = (double)k;
+                init = true;
+                break;
+            }
+        }
+        if (!init) {
+            maxval = sp[0];
+        }
+        for (ompIndexType k = 0; k < (ompIndexType)elementCount; ++k) {
+            T val = sp[k];
+            if (!std::isnan(val)) {
+                if (val > maxval) {
+                    maxval = val;
+                }
+            }
+        }
+        if (init) {
+            dp[0] = maxval;
+        } else {
+            dp[0] = (T)std::nan("NaN");
+        }
+    } else {
+        T maxval;
+        maxval = sp[0];
+        for (ompIndexType k = 0; k < (ompIndexType)elementCount; ++k) {
+            T val = sp[k];
+            if (val > maxval || std::isnan(val)) {
+                maxval = val;
             }
         }
         dp[0] = maxval;
-        if (iptr != nullptr) {
-            iptr[0] = maxdex + 1;
+    }
+}
+//=============================================================================
+template <class T>
+void
+TMaxReal(bool omitnan, const T* sp, T* dp, double* iptr, indexType planes, indexType planesize,
+    indexType linesize)
+{
+    if (omitnan) {
+        T maxval;
+        for (ompIndexType i = 0; i < (ompIndexType)planes; i++) {
+            for (indexType j = 0; j < planesize; j++) {
+                bool init = false;
+                double maxdex = 0;
+                for (indexType k = 0; k < linesize; k++) {
+                    T val = sp[i * planesize * linesize + j + k * planesize];
+                    if (!std::isnan(val)) {
+                        maxval = val;
+                        maxdex = (double)k;
+                        init = true;
+                        break;
+                    }
+                }
+                if (init) {
+                    for (ompIndexType k = 0; k < (ompIndexType)linesize; k++) {
+                        T val = sp[i * planesize * linesize + j + k * planesize];
+                        if (!std::isnan(val)) {
+                            if (val > maxval) {
+                                maxval = val;
+                                maxdex = (double)k;
+                            }
+                        }
+                    }
+                }
+                if (init) {
+                    dp[i * planesize + j] = maxval;
+                    if (iptr != nullptr) {
+                        iptr[i * planesize + j] = maxdex + 1;
+                    }
+                } else {
+                    dp[i * planesize + j] = (T)std::nan("NaN");
+                    if (iptr != nullptr) {
+                        iptr[i * planesize + j] = 0;
+                    }
+                }
+            }
         }
     } else {
+        T maxval;
+        for (ompIndexType i = 0; i < (ompIndexType)planes; i++) {
+            for (indexType j = 0; j < planesize; j++) {
+                double maxdex = 0;
+                maxval = sp[i * planesize * linesize + j + 0 * planesize];
+                maxdex = (double)0;
+                for (ompIndexType k = 0; k < (ompIndexType)linesize; k++) {
+                    T val = sp[i * planesize * linesize + j + k * planesize];
+                    if (val > maxval || std::isnan(val)) {
+
+                        maxval = val;
+                        maxdex = (double)k;
+                    }
+                }
+                dp[i * planesize + j] = maxval;
+                if (iptr != nullptr) {
+                    iptr[i * planesize + j] = maxdex + 1;
+                }
+            }
+        }
+    }
+}
+//=============================================================================
+template <class T>
+void
+TMaxAllComplex(bool omitnan, const T* sp, T* dp, indexType elementCount)
+{
+    if (omitnan) {
+        T maxval;
+        T maxval_r;
+        T maxval_i;
+        T tstval;
+        bool init = false;
+        for (indexType k = 0; k < elementCount; ++k) {
+            T val_re = sp[2 * k];
+            T val_im = sp[(2 * k) + 1];
+            if ((!std::isnan(val_re)) && (!std::isnan(val_im))) {
+                tstval = complex_abs(val_re, val_im);
+                init = true;
+                break;
+            }
+        }
+
+        if (init) {
+            for (ompIndexType k = 0; k < (ompIndexType)elementCount; ++k) {
+                T val_re = sp[2 * k];
+                T val_im = sp[(2 * k) + 1];
+                if ((!std::isnan(val_re)) && (!std::isnan(val_im))) {
+                    tstval = complex_abs(val_re, val_im);
+                    if (tstval > maxval) {
+                        maxval = tstval;
+                        maxval_r = val_re;
+                        maxval_i = val_im;
+                    }
+                }
+            }
+        }
+
+        if (init) {
+            dp[0] = maxval_r;
+            dp[1] = maxval_i;
+        } else {
+            dp[0] = (T)atof("nan");
+            dp[1] = (T)atof("nan");
+        }
+
+    } else {
+        T maxval;
+        T maxval_r;
+        T maxval_i;
+        T tstval;
+        tstval = complex_abs(sp[0], sp[1]);
+        for (ompIndexType k = 1; k < (ompIndexType)elementCount; ++k) {
+            T val_re = sp[2 * k];
+            T val_im = sp[(2 * k) + 1];
+            tstval = complex_abs(val_re, val_im);
+            if (tstval > maxval || std::isnan(tstval)) {
+                maxval = tstval;
+                maxval_r = val_re;
+                maxval_i = val_im;
+            }
+        }
+        dp[0] = maxval_r;
+        dp[1] = maxval_i;
+    }
+}
+//=============================================================================
+template <class T>
+void
+TMaxComplex(bool omitnan, const T* sp, T* dp, double* iptr, indexType planes, indexType planesize,
+    indexType linesize)
+{
+    if (omitnan) {
+        T maxval;
+        T maxval_r;
+        T maxval_i;
+        T tstval;
+        double maxdex;
+        bool init = false;
+        for (ompIndexType i = 0; i < (ompIndexType)planes; i++) {
+            for (indexType j = 0; j < planesize; j++) {
+                init = false;
+                maxdex = 0;
+                for (indexType k = 0; k < linesize; k++) {
+                    T val_re = sp[2 * (i * planesize * linesize + j + k * planesize)];
+                    T val_im = sp[2 * (i * planesize * linesize + j + k * planesize) + 1];
+                    if ((!std::isnan(val_re)) && (!std::isnan(val_im))) {
+                        tstval = complex_abs(val_re, val_im);
+                        init = true;
+                        maxval = tstval;
+                        maxdex = (double)k;
+                        maxval_r = val_re;
+                        maxval_i = val_im;
+                        break;
+                    }
+                }
+                if (init) {
+                    for (ompIndexType k = 0; k < (ompIndexType)linesize; k++) {
+                        T val_re = sp[2 * (i * planesize * linesize + j + k * planesize)];
+                        T val_im = sp[2 * (i * planesize * linesize + j + k * planesize) + 1];
+                        if ((!std::isnan(val_re)) && (!std::isnan(val_im))) {
+                            tstval = complex_abs(val_re, val_im);
+                            if (tstval > maxval) {
+                                maxval = tstval;
+                                maxdex = (double)k;
+                                maxval_r = val_re;
+                                maxval_i = val_im;
+                            }
+                        }
+                    }
+                }
+                if (init) {
+                    dp[2 * (i * planesize + j)] = maxval_r;
+                    dp[2 * (i * planesize + j) + 1] = maxval_i;
+                    if (iptr != nullptr) {
+                        iptr[i * planesize + j] = maxdex + 1;
+                    }
+                } else {
+                    dp[2 * (i * planesize + j)] = (T)std::nan("NaN");
+                    dp[2 * (i * planesize + j) + 1] = (T)std::nan("NaN");
+                    if (iptr != nullptr) {
+                        iptr[i * planesize + j] = 0;
+                    }
+                }
+            }
+        }
+
+    } else {
+        T maxval;
+        T maxval_r;
+        T maxval_i;
+        T tstval;
+        double maxdex;
         for (indexType i = 0; i < planes; i++) {
             for (indexType j = 0; j < planesize; j++) {
-                maxval = sp[i * planesize * linesize + j];
+                maxval_r = sp[2 * (i * planesize * linesize + j + 0 * planesize)];
+                maxval_i = sp[2 * (i * planesize * linesize + j + 0 * planesize) + 1];
+                maxval = complex_abs(maxval_r, maxval_i);
                 maxdex = 0;
-#if defined(_NLS_WITH_OPENMP)
-#pragma omp parallel for ordered
-#endif
+
                 for (ompIndexType k = 0; k < (ompIndexType)linesize; k++) {
-                    if (sp[i * planesize * linesize + j + k * planesize] > maxval) {
-                        maxval = sp[i * planesize * linesize + j + k * planesize];
-                        maxdex = (double)k;
-                    }
-                }
-                dp[i * planesize + j] = maxval;
-                if (iptr != nullptr) {
-                    iptr[i * planesize + j] = maxdex + 1;
-                }
-            }
-        }
-    }
-}
-//=============================================================================
-template <class T>
-void
-TMaxReal(
-    const T* sp, T* dp, double* iptr, indexType planes, indexType planesize, indexType linesize)
-{
-    T maxval;
-    double maxdex = 0;
-    bool init = false;
-#if defined(_NLS_WITH_OPENMP)
-#pragma omp parallel for ordered
-#endif
-    for (ompIndexType i = 0; i < (ompIndexType)planes; i++) {
-        for (indexType j = 0; j < planesize; j++) {
-            init = false;
-            maxdex = 0;
-            for (ompIndexType k = 0; k < (ompIndexType)linesize; k++) {
-                if (!std::isnan(sp[i * planesize * linesize + j + k * planesize])) {
-                    if (!init) {
-                        init = true;
-                        maxval = sp[i * planesize * linesize + j + k * planesize];
-                        maxdex = (double)k;
-                    } else if (sp[i * planesize * linesize + j + k * planesize] > maxval) {
-                        maxval = sp[i * planesize * linesize + j + k * planesize];
-                        maxdex = (double)k;
-                    }
-                }
-            }
-            if (init) {
-                dp[i * planesize + j] = maxval;
-                if (iptr != nullptr) {
-                    iptr[i * planesize + j] = maxdex + 1;
-                }
-            } else {
-                dp[i * planesize + j] = (T)atof("nan");
-                if (iptr != nullptr) {
-                    iptr[i * planesize + j] = 0;
-                }
-            }
-        }
-    }
-}
-//=============================================================================
-template <class T>
-void
-TMaxComplex(
-    const T* sp, T* dp, double* iptr, indexType planes, indexType planesize, indexType linesize)
-{
-    T maxval;
-    T maxval_r;
-    T maxval_i;
-    T tstval;
-    double maxdex;
-    bool init = false;
-#if defined(_NLS_WITH_OPENMP)
-#pragma omp parallel for ordered
-#endif
-    for (ompIndexType i = 0; i < (ompIndexType)planes; i++) {
-        for (indexType j = 0; j < planesize; j++) {
-            init = false;
-            maxdex = 0;
-            for (ompIndexType k = 0; k < (ompIndexType)linesize; k++) {
-                if ((!std::isnan(sp[2 * (i * planesize * linesize + j + k * planesize)]))
-                    && (!std::isnan(sp[2 * (i * planesize * linesize + j + k * planesize) + 1]))) {
-                    tstval = complex_abs(sp[2 * (i * planesize * linesize + j + k * planesize)],
-                        sp[2 * (i * planesize * linesize + j + k * planesize) + 1]);
-                    if (!init) {
-                        init = true;
+                    T val_re = sp[2 * (i * planesize * linesize + j + k * planesize)];
+                    T val_im = sp[2 * (i * planesize * linesize + j + k * planesize) + 1];
+                    tstval = complex_abs(val_re, val_im);
+                    if (tstval > maxval || std::isnan(tstval)) {
                         maxval = tstval;
                         maxdex = (double)k;
-                        maxval_r = sp[2 * (i * planesize * linesize + j + k * planesize)];
-                        maxval_i = sp[2 * (i * planesize * linesize + j + k * planesize) + 1];
-                    } else if (tstval > maxval) {
-                        maxval = tstval;
-                        maxdex = (double)k;
-                        maxval_r = sp[2 * (i * planesize * linesize + j + k * planesize)];
-                        maxval_i = sp[2 * (i * planesize * linesize + j + k * planesize) + 1];
+                        maxval_r = val_re;
+                        maxval_i = val_im;
                     }
                 }
-            }
-            if (init) {
                 dp[2 * (i * planesize + j)] = maxval_r;
                 dp[2 * (i * planesize + j) + 1] = maxval_i;
                 if (iptr != nullptr) {
                     iptr[i * planesize + j] = maxdex + 1;
-                }
-            } else {
-                dp[2 * (i * planesize + j)] = (T)atof("nan");
-                dp[2 * (i * planesize + j) + 1] = (T)atof("nan");
-                if (iptr != nullptr) {
-                    iptr[i * planesize + j] = 0;
                 }
             }
         }
@@ -209,7 +463,7 @@ TMaxComplex(
 //=============================================================================
 // C = max(A, B)
 ArrayOf
-Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
+Maximum(bool omitNaN, const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
 {
     needToOverload = false;
     if (A.isEmpty()) {
@@ -268,7 +522,7 @@ Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
         _B.promoteType(outType);
         logical* ptr = (logical*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterReal<logical>((const logical*)_A.getDataPointer(),
+        TMaxLessInteger<logical>((const logical*)_A.getDataPointer(),
             (const logical*)_B.getDataPointer(), ptr, outDim.getElementCount(), AStride, BStride);
     } break;
     case NLS_UINT8: {
@@ -276,7 +530,7 @@ Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
         _B.promoteType(outType);
         uint8* ptr = (uint8*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterReal<uint8>((const uint8*)_A.getDataPointer(), (const uint8*)_B.getDataPointer(),
+        TMaxLessInteger<uint8>((const uint8*)_A.getDataPointer(), (const uint8*)_B.getDataPointer(),
             ptr, outDim.getElementCount(), AStride, BStride);
     } break;
     case NLS_INT8: {
@@ -284,7 +538,7 @@ Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
         _B.promoteType(outType);
         int8* ptr = (int8*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterReal<int8>((const int8*)_A.getDataPointer(), (const int8*)_B.getDataPointer(),
+        TMaxLessInteger<int8>((const int8*)_A.getDataPointer(), (const int8*)_B.getDataPointer(),
             ptr, outDim.getElementCount(), AStride, BStride);
     } break;
     case NLS_UINT16: {
@@ -292,7 +546,7 @@ Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
         _B.promoteType(outType);
         uint16* ptr = (uint16*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterReal<uint16>((const uint16*)_A.getDataPointer(),
+        TMaxLessInteger<uint16>((const uint16*)_A.getDataPointer(),
             (const uint16*)_B.getDataPointer(), ptr, outDim.getElementCount(), AStride, BStride);
     } break;
     case NLS_INT16: {
@@ -300,7 +554,7 @@ Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
         _B.promoteType(outType);
         int16* ptr = (int16*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterReal<int16>((const int16*)_A.getDataPointer(), (const int16*)_B.getDataPointer(),
+        TMaxLessInteger<int16>((const int16*)_A.getDataPointer(), (const int16*)_B.getDataPointer(),
             ptr, outDim.getElementCount(), AStride, BStride);
     } break;
     case NLS_UINT32: {
@@ -308,7 +562,7 @@ Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
         _B.promoteType(outType);
         uint32* ptr = (uint32*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterReal<uint32>((const uint32*)_A.getDataPointer(),
+        TMaxLessInteger<uint32>((const uint32*)_A.getDataPointer(),
             (const uint32*)_B.getDataPointer(), ptr, outDim.getElementCount(), AStride, BStride);
     } break;
     case NLS_INT32: {
@@ -316,7 +570,7 @@ Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
         _B.promoteType(outType);
         int32* ptr = (int32*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterReal<int32>((const int32*)_A.getDataPointer(), (const int32*)_B.getDataPointer(),
+        TMaxLessInteger<int32>((const int32*)_A.getDataPointer(), (const int32*)_B.getDataPointer(),
             ptr, outDim.getElementCount(), AStride, BStride);
     } break;
     case NLS_UINT64: {
@@ -324,7 +578,7 @@ Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
         _B.promoteType(outType);
         uint64* ptr = (uint64*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterReal<uint64>((const uint64*)_A.getDataPointer(),
+        TMaxLessInteger<uint64>((const uint64*)_A.getDataPointer(),
             (const uint64*)_B.getDataPointer(), ptr, outDim.getElementCount(), AStride, BStride);
     } break;
     case NLS_INT64: {
@@ -332,7 +586,7 @@ Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
         _B.promoteType(outType);
         int64* ptr = (int64*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterReal<int64>((const int64*)_A.getDataPointer(), (const int64*)_B.getDataPointer(),
+        TMaxLessInteger<int64>((const int64*)_A.getDataPointer(), (const int64*)_B.getDataPointer(),
             ptr, outDim.getElementCount(), AStride, BStride);
     } break;
     case NLS_SINGLE: {
@@ -340,7 +594,7 @@ Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
         _B.promoteType(outType);
         single* ptr = (single*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterReal<single>((const single*)_A.getDataPointer(),
+        TMaxLessReal<single>(omitNaN, (const single*)_A.getDataPointer(),
             (const single*)_B.getDataPointer(), ptr, outDim.getElementCount(), AStride, BStride);
     } break;
     case NLS_DOUBLE: {
@@ -348,7 +602,7 @@ Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
         _B.promoteType(outType);
         double* ptr = (double*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterReal<double>((const double*)_A.getDataPointer(),
+        TMaxLessReal<double>(omitNaN, (const double*)_A.getDataPointer(),
             (const double*)_B.getDataPointer(), ptr, outDim.getElementCount(), AStride, BStride);
     } break;
     case NLS_SCOMPLEX: {
@@ -356,23 +610,29 @@ Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
         _B.promoteType(outType);
         single* ptr = (single*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterComplex<single>((const single*)_A.getDataPointer(),
+        TMaxLessComplex<single>(omitNaN, (const single*)_A.getDataPointer(),
             (const single*)_B.getDataPointer(), ptr, outDim.getElementCount(), AStride, BStride);
+        if (res.allReal()) {
+            res.promoteType(NLS_SINGLE);
+        }
     } break;
     case NLS_DCOMPLEX: {
         _A.promoteType(outType);
         _B.promoteType(outType);
         double* ptr = (double*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterComplex<double>((const double*)_A.getDataPointer(),
+        TMaxLessComplex<double>(omitNaN, (const double*)_A.getDataPointer(),
             (const double*)_B.getDataPointer(), ptr, outDim.getElementCount(), AStride, BStride);
+        if (res.allReal()) {
+            res.promoteType(NLS_DOUBLE);
+        }
     } break;
     case NLS_CHAR: {
         _A.promoteType(outType);
         _B.promoteType(outType);
         charType* ptr = (charType*)ArrayOf::allocateArrayOf(outType, outDim.getElementCount());
         res = ArrayOf(outType, outDim, ptr);
-        TMaxGreaterReal<charType>((const charType*)_A.getDataPointer(),
+        TMaxLessInteger<charType>((const charType*)_A.getDataPointer(),
             (const charType*)_B.getDataPointer(), ptr, outDim.getElementCount(), AStride, BStride);
     } break;
     case NLS_HANDLE:
@@ -388,7 +648,7 @@ Maximum(const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
 //=============================================================================
 // [M, i] = max(A)
 ArrayOfVector
-Maximum(const ArrayOf& A, int nLhs, bool& needToOverload)
+Maximum(bool omitNaN, const ArrayOf& A, int nLhs, bool& needToOverload)
 {
     ArrayOfVector retval;
     needToOverload = false;
@@ -409,14 +669,14 @@ Maximum(const ArrayOf& A, int nLhs, bool& needToOverload)
             d++;
         }
         indexType dim = d + 1;
-        retval = Maximum(A, dim, nLhs, needToOverload);
+        retval = Maximum(omitNaN, A, dim, nLhs, needToOverload);
     }
     return retval;
 }
 //=============================================================================
 // [M, i] = max(A, [], dim)
 ArrayOfVector
-Maximum(const ArrayOf& A, indexType dim, int nLhs, bool& needToOverload)
+Maximum(bool omitNaN, const ArrayOf& A, indexType dim, int nLhs, bool& needToOverload)
 {
     ArrayOfVector retval;
     needToOverload = false;
@@ -459,6 +719,24 @@ Maximum(const ArrayOf& A, indexType dim, int nLhs, bool& needToOverload)
     ArrayOf index;
     double* iptr = nullptr;
     ArrayOf res;
+
+    if (dimsA.equals(outDim)) {
+        outDim.simplify();
+        retval.push_back(A);
+        if (nLhs > 1) {
+            iptr = (double*)ArrayOf::allocateArrayOf(NLS_DOUBLE, outDim.getElementCount());
+            index = ArrayOf(NLS_DOUBLE, outDim, iptr);
+#if defined(_NLS_WITH_OPENMP)
+#pragma omp parallel for
+#endif
+            for (ompIndexType k = 0; k < (ompIndexType)outDim.getElementCount(); ++k) {
+                iptr[k] = 1;
+            }
+            retval.push_back(index);
+        }
+        return retval;
+    }
+
     switch (A.getDataClass()) {
     case NLS_HANDLE:
     case NLS_CELL_ARRAY:
@@ -563,8 +841,8 @@ Maximum(const ArrayOf& A, indexType dim, int nLhs, bool& needToOverload)
             index = ArrayOf(NLS_DOUBLE, outDim, iptr);
         }
         single* ptr = (single*)ArrayOf::allocateArrayOf(NLS_SINGLE, outDim.getElementCount());
-        TMaxReal<single>(
-            (const single*)A.getDataPointer(), (single*)ptr, iptr, planecount, planesize, linesize);
+        TMaxReal<single>(omitNaN, (const single*)A.getDataPointer(), (single*)ptr, iptr, planecount,
+            planesize, linesize);
         res = ArrayOf(NLS_SINGLE, outDim, ptr);
     } break;
     case NLS_DOUBLE: {
@@ -573,8 +851,8 @@ Maximum(const ArrayOf& A, indexType dim, int nLhs, bool& needToOverload)
             index = ArrayOf(NLS_DOUBLE, outDim, iptr);
         }
         double* ptr = (double*)ArrayOf::allocateArrayOf(NLS_DOUBLE, outDim.getElementCount());
-        TMaxReal<double>(
-            (const double*)A.getDataPointer(), (double*)ptr, iptr, planecount, planesize, linesize);
+        TMaxReal<double>(omitNaN, (const double*)A.getDataPointer(), (double*)ptr, iptr, planecount,
+            planesize, linesize);
         res = ArrayOf(NLS_DOUBLE, outDim, ptr);
     } break;
     case NLS_SCOMPLEX: {
@@ -583,9 +861,12 @@ Maximum(const ArrayOf& A, indexType dim, int nLhs, bool& needToOverload)
             index = ArrayOf(NLS_DOUBLE, outDim, iptr);
         }
         single* ptr = (single*)ArrayOf::allocateArrayOf(NLS_SCOMPLEX, outDim.getElementCount());
-        TMaxComplex<single>(
-            (const single*)A.getDataPointer(), (single*)ptr, iptr, planecount, planesize, linesize);
+        TMaxComplex<single>(omitNaN, (const single*)A.getDataPointer(), (single*)ptr, iptr,
+            planecount, planesize, linesize);
         res = ArrayOf(NLS_SCOMPLEX, outDim, ptr);
+        if (res.allReal()) {
+            res.promoteType(NLS_SINGLE);
+        }
     } break;
     case NLS_DCOMPLEX: {
         if (nLhs > 1) {
@@ -593,9 +874,12 @@ Maximum(const ArrayOf& A, indexType dim, int nLhs, bool& needToOverload)
             index = ArrayOf(NLS_DOUBLE, outDim, iptr);
         }
         double* ptr = (double*)ArrayOf::allocateArrayOf(NLS_DCOMPLEX, outDim.getElementCount());
-        TMaxComplex<double>(
-            (const double*)A.getDataPointer(), (double*)ptr, iptr, planecount, planesize, linesize);
+        TMaxComplex<double>(omitNaN, (const double*)A.getDataPointer(), (double*)ptr, iptr,
+            planecount, planesize, linesize);
         res = ArrayOf(NLS_DCOMPLEX, outDim, ptr);
+        if (res.allReal()) {
+            res.promoteType(NLS_DOUBLE);
+        }
     } break;
     case NLS_CHAR: {
         if (nLhs > 1) {
@@ -615,6 +899,116 @@ Maximum(const ArrayOf& A, indexType dim, int nLhs, bool& needToOverload)
         retval.push_back(index);
     }
     return retval;
+}
+//=============================================================================
+ArrayOf
+MaximumAll(bool omitNaN, const ArrayOf& A, bool& needToOverload)
+{
+    ArrayOf res;
+    needToOverload = false;
+    if (A.isSparse()) {
+        needToOverload = true;
+        return res;
+    }
+    if (A.isEmpty()) {
+        return A;
+    }
+    if (A.isScalar()) {
+        return A;
+    }
+    Dimensions outDim(1, 1);
+    indexType elementCount = A.getDimensions().getElementCount();
+    switch (A.getDataClass()) {
+    case NLS_HANDLE:
+    case NLS_CELL_ARRAY:
+    case NLS_STRUCT_ARRAY:
+    case NLS_STRING_ARRAY:
+    default: {
+        needToOverload = true;
+    } break;
+
+    case NLS_LOGICAL: {
+        logical* ptr = (logical*)ArrayOf::allocateArrayOf(NLS_LOGICAL, outDim.getElementCount());
+        TMaxAllInteger<logical>((const logical*)A.getDataPointer(), (logical*)ptr, elementCount);
+        res = ArrayOf(NLS_LOGICAL, outDim, ptr);
+    } break;
+    case NLS_UINT8: {
+        uint8* ptr = (uint8*)ArrayOf::allocateArrayOf(NLS_UINT8, outDim.getElementCount());
+        TMaxAllInteger<uint8>((const uint8*)A.getDataPointer(), (uint8*)ptr, elementCount);
+        res = ArrayOf(NLS_UINT8, outDim, ptr);
+    } break;
+    case NLS_INT8: {
+        int8* ptr = (int8*)ArrayOf::allocateArrayOf(NLS_INT8, outDim.getElementCount());
+        TMaxAllInteger<int8>((const int8*)A.getDataPointer(), (int8*)ptr, elementCount);
+        res = ArrayOf(NLS_INT8, outDim, ptr);
+    } break;
+    case NLS_UINT16: {
+        uint16* ptr = (uint16*)ArrayOf::allocateArrayOf(NLS_UINT16, outDim.getElementCount());
+        TMaxAllInteger<uint16>((const uint16*)A.getDataPointer(), (uint16*)ptr, elementCount);
+        res = ArrayOf(NLS_UINT16, outDim, ptr);
+    } break;
+    case NLS_INT16: {
+        int16* ptr = (int16*)ArrayOf::allocateArrayOf(NLS_INT16, outDim.getElementCount());
+        TMaxAllInteger<int16>((const int16*)A.getDataPointer(), (int16*)ptr, elementCount);
+        res = ArrayOf(NLS_INT16, outDim, ptr);
+    } break;
+    case NLS_UINT32: {
+        uint32* ptr = (uint32*)ArrayOf::allocateArrayOf(NLS_UINT32, outDim.getElementCount());
+        TMaxAllInteger<uint32>((const uint32*)A.getDataPointer(), (uint32*)ptr, elementCount);
+        res = ArrayOf(NLS_UINT32, outDim, ptr);
+    } break;
+    case NLS_INT32: {
+        int32* ptr = (int32*)ArrayOf::allocateArrayOf(NLS_INT32, outDim.getElementCount());
+        TMaxAllInteger<int32>((const int32*)A.getDataPointer(), (int32*)ptr, elementCount);
+        res = ArrayOf(NLS_INT32, outDim, ptr);
+    } break;
+    case NLS_UINT64: {
+        uint64* ptr = (uint64*)ArrayOf::allocateArrayOf(NLS_UINT64, outDim.getElementCount());
+        TMaxAllInteger<uint64>((const uint64*)A.getDataPointer(), (uint64*)ptr, elementCount);
+        res = ArrayOf(NLS_UINT64, outDim, ptr);
+    } break;
+    case NLS_INT64: {
+        int64* ptr = (int64*)ArrayOf::allocateArrayOf(NLS_INT64, outDim.getElementCount());
+        TMaxAllInteger<int64>((const int64*)A.getDataPointer(), (int64*)ptr, elementCount);
+        res = ArrayOf(NLS_INT64, outDim, ptr);
+    } break;
+    case NLS_SINGLE: {
+        single* ptr = (single*)ArrayOf::allocateArrayOf(NLS_SINGLE, outDim.getElementCount());
+        TMaxAllReal<single>(omitNaN, (const single*)A.getDataPointer(), (single*)ptr, elementCount);
+        res = ArrayOf(NLS_SINGLE, outDim, ptr);
+    } break;
+    case NLS_DOUBLE: {
+        double* ptr = (double*)ArrayOf::allocateArrayOf(NLS_DOUBLE, outDim.getElementCount());
+        TMaxAllReal<double>(omitNaN, (const double*)A.getDataPointer(), (double*)ptr, elementCount);
+        res = ArrayOf(NLS_DOUBLE, outDim, ptr);
+    } break;
+    case NLS_SCOMPLEX: {
+        single* ptr = (single*)ArrayOf::allocateArrayOf(NLS_SCOMPLEX, outDim.getElementCount());
+        TMaxAllComplex<single>(
+            omitNaN, (const single*)A.getDataPointer(), (single*)ptr, elementCount);
+        res = ArrayOf(NLS_SCOMPLEX, outDim, ptr);
+        if (res.allReal()) {
+            res.promoteType(NLS_SINGLE);
+        }
+    } break;
+    case NLS_DCOMPLEX: {
+        double* ptr = (double*)ArrayOf::allocateArrayOf(NLS_DCOMPLEX, outDim.getElementCount());
+        TMaxAllComplex<double>(
+            omitNaN, (const double*)A.getDataPointer(), (double*)ptr, elementCount);
+        res = ArrayOf(NLS_DCOMPLEX, outDim, ptr);
+        if (res.allReal()) {
+            res.promoteType(NLS_DOUBLE);
+        }
+    } break;
+    case NLS_CHAR: {
+        ArrayOf _A(A);
+        _A.promoteType(NLS_DOUBLE);
+        double* ptr = (double*)ArrayOf::allocateArrayOf(NLS_DOUBLE, outDim.getElementCount());
+        TMaxAllInteger<double>((const double*)_A.getDataPointer(), (double*)ptr, elementCount);
+        res = ArrayOf(NLS_DOUBLE, outDim, ptr);
+    } break;
+    }
+    return res;
 }
 //=============================================================================
 }