+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.4.2 (2020-02-XX)
 
-* min, max builtin: Minimum/Maximum elements of an array
+* min, max builtin: Minimum/Maximum elements of an array with nanflag and 'all' support.
 
 * flipud: Flip array up to down.
 
@@ -8,7 +8,7 @@
 
 * flip: Flip order of elements.
 
-* flipdim: Flip array along specified dimension?
+* flipdim: Flip array along specified dimension.
 
 * log2 builtin: Base 2 logarithm and floating-point number dissection.